jeez it's 12:54 AM... I need to head home

+++ core/src/com/rockbite/hackathon/sm/components/CardComponent.java
@@ -7,7 +7,7 @@ import org.json.JSONObject;
 
 public class CardComponent implements Component {
 
-    public long playerId;
+    public int playerId;
 
     public int slot;
